#31, #32, #34: Implement Array class to allocate cpu/gpu byte buffer

See merge request TensorRT/poc/tripy!61

+++ tests/flat_ir/test_flat_ir.py
@@ -89,8 +89,8 @@ class TestFlatIR:
             str(flat_ir)
             == dedent(
                 """
-                t0 : data=([0.]), shape=((1,)), dtype=(float32), stride=(), loc=(cpu:0)
-                t1 : data=([1.]), shape=((1,)), dtype=(float32), stride=(), loc=(cpu:0)
+                t0 : data=([0]), shape=((1,)), dtype=(int32), stride=(), loc=(cpu:0)
+                t1 : data=([1]), shape=((1,)), dtype=(int32), stride=(), loc=(cpu:0)
                 t2 = t0 + t1
                 outputs: t2
                 """
@@ -112,8 +112,8 @@ class TestFlatIR:
 
     def test_multiple_outputs(self):
         shape = 1
-        a = Tensor(np.ones(shape))
-        b = Tensor(np.ones(shape))
+        a = Tensor(np.ones(shape, dtype=np.float32))
+        b = Tensor(np.ones(shape, dtype=np.float32))
 
         c = a + b
         d = c + c
@@ -125,8 +125,8 @@ class TestFlatIR:
             str(flat_ir)
             == dedent(
                 """
-                t0 : data=([1.]), shape=((1,)), dtype=(float32), stride=(), loc=(cpu:0)
-                t1 : data=([1.]), shape=((1,)), dtype=(float32), stride=(), loc=(cpu:0)
+                t0 : data=([1.0]), shape=((1,)), dtype=(float32), stride=(), loc=(cpu:0)
+                t1 : data=([1.0]), shape=((1,)), dtype=(float32), stride=(), loc=(cpu:0)
                 t2 = t0 + t1
                 t3 = t2 + t2
                 outputs: t2, t3
@@ -146,8 +146,10 @@ class TestFlatIR:
 
     def test_all_inputs(self):
         shape = 1
-        a = Tensor(np.ones(shape))
-        b = Tensor(np.ones(shape))
+        # Need explicit data type here since by default dtype is np.float64 which is not yet supported.
+        # (38): Add cast operation to support unsupported backend types.
+        a = Tensor(np.ones(shape, dtype=np.float32))
+        b = Tensor(np.ones(shape, dtype=np.float32))
         # a and b are inputs
         a.const_fold = False
         b.const_fold = False
@@ -170,8 +172,8 @@ class TestFlatIR:
 
     def test_const_and_input(self):
         shape = 1
-        a = Tensor(np.ones(shape))
-        b = Tensor(np.ones(shape))
+        a = Tensor(np.ones(shape, dtype=np.float32))
+        b = Tensor(np.ones(shape, dtype=np.float32))
         # a is an input
         a.const_fold = False
 
@@ -184,7 +186,7 @@ class TestFlatIR:
                 """
                 inputs:
                     t0 : shape=((1,)), dtype=(float32)
-                t1 : data=([1.]), shape=((1,)), dtype=(float32), stride=(), loc=(cpu:0)
+                t1 : data=([1.0]), shape=((1,)), dtype=(float32), stride=(), loc=(cpu:0)
                 t2 = t0 + t1
                 outputs: t2
                 """
